+++ example/file.hpp
@@ -10,7 +10,7 @@
 #ifndef BOOST_JSON_EXAMPLE_FILE_HPP
 #define BOOST_JSON_EXAMPLE_FILE_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 #include <cstdio>
 #include <string>
