+++ include/boost/json.hpp
@@ -10,7 +10,7 @@
 #ifndef BOOST_JSON_HPP
 #define BOOST_JSON_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 
 #include <boost/json/array.hpp>
 #include <boost/json/error.hpp>
