+++ include/boost/json/array.hpp
@@ -10,7 +10,7 @@
 #ifndef BOOST_JSON_ARRAY_HPP
 #define BOOST_JSON_ARRAY_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 #include <boost/json/kind.hpp>
 #include <boost/json/storage_ptr.hpp>
 #include <boost/json/detail/array_impl.hpp>
