+++ include/boost/json/detail/array_impl.hpp
@@ -10,7 +10,7 @@
 #ifndef BOOST_JSON_DETAIL_ARRAY_IMPL_HPP
 #define BOOST_JSON_DETAIL_ARRAY_IMPL_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 #include <boost/json/storage_ptr.hpp>
 #include <cstdint>
 #include <cstring>
