+++ include/boost/json/detail/buffer.hpp
@@ -10,7 +10,8 @@
 #ifndef BOOST_JSON_DETAIL_BUFFER_HPP
 #define BOOST_JSON_DETAIL_BUFFER_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
+#include <boost/json/string_view.hpp>
 #include <cstring>
 
 namespace boost {
