+++ include/boost/json/detail/object_impl.hpp
@@ -10,8 +10,9 @@
 #ifndef BOOST_JSON_DETAIL_OBJECT_IMPL_HPP
 #define BOOST_JSON_DETAIL_OBJECT_IMPL_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 #include <boost/json/storage_ptr.hpp>
+#include <boost/json/string_view.hpp>
 #include <cstddef>
 #include <cstring>
 #include <type_traits>
