+++ include/boost/json/detail/ryu/detail/d2s.hpp
@@ -22,7 +22,7 @@
 #ifndef BOOST_JSON_DETAIL_RYU_DETAIL_D2S_HPP
 #define BOOST_JSON_DETAIL_RYU_DETAIL_D2S_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 #include <boost/json/detail/ryu/detail/common.hpp>
 
 // Only include the full table if we're not optimizing for size.
