+++ include/boost/json/detail/ryu/ryu.hpp
@@ -22,7 +22,7 @@
 #ifndef BOOST_JSON_DETAIL_RYU_HPP
 #define BOOST_JSON_DETAIL_RYU_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 
 namespace boost {
 namespace json {
