+++ include/boost/json/detail/string_impl.hpp
@@ -11,7 +11,7 @@
 #ifndef BOOST_JSON_DETAIL_STRING_IMPL_HPP
 #define BOOST_JSON_DETAIL_STRING_IMPL_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 #include <boost/json/kind.hpp>
 #include <boost/json/storage_ptr.hpp>
 #include <algorithm>
