+++ include/boost/json/except.hpp
@@ -10,7 +10,7 @@
 #ifndef BOOST_JSON_EXCEPT_HPP
 #define BOOST_JSON_EXCEPT_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 #include <stdexcept>
 
 namespace boost {
