+++ include/boost/json/monotonic_resource.hpp
@@ -10,7 +10,7 @@
 #ifndef BOOST_JSON_MONOTONIC_RESOURCE_HPP
 #define BOOST_JSON_MONOTONIC_RESOURCE_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 #include <boost/json/storage_ptr.hpp>
 
 namespace boost {
