+++ include/boost/json/parser.hpp
@@ -10,7 +10,7 @@
 #ifndef BOOST_JSON_PARSER_HPP
 #define BOOST_JSON_PARSER_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 #include <boost/json/storage_ptr.hpp>
 #include <boost/json/value.hpp>
 #include <boost/json/string.hpp>
