+++ include/boost/json/storage_ptr.hpp
@@ -10,7 +10,7 @@
 #ifndef BOOST_JSON_STORAGE_PTR_HPP
 #define BOOST_JSON_STORAGE_PTR_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 #include <boost/json/memory_resource.hpp>
 #include <boost/json/detail/counted_resource.hpp>
 #include <boost/json/detail/default_resource.hpp>
