+++ include/boost/json/string.hpp
@@ -11,8 +11,9 @@
 #ifndef BOOST_JSON_STRING_HPP
 #define BOOST_JSON_STRING_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 #include <boost/json/storage_ptr.hpp>
+#include <boost/json/string_view.hpp>
 #include <boost/json/detail/string_impl.hpp>
 #include <boost/pilfer.hpp>
 #include <algorithm>
