+++ include/boost/json/value_ref.hpp
@@ -10,7 +10,7 @@
 #ifndef BOOST_JSON_VALUE_REF_HPP
 #define BOOST_JSON_VALUE_REF_HPP
 
-#include <boost/json/config.hpp>
+#include <boost/json/detail/config.hpp>
 #include <boost/json/storage_ptr.hpp>
 #include <boost/json/string.hpp>
 #include <initializer_list>
