+++ test/string_view.cpp
@@ -1,0 +1,11 @@
+//
+// Copyright (c) 2019 Vinnie Falco (vinnie.falco@gmail.com)
+//
+// Distributed under the Boost Software License, Version 1.0. (See accompanying
+// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+//
+// Official repository: https://github.com/vinniefalco/BeastLounge
+//
+
+// Test that header file is self-contained.
+#include <boost/json/string_view.hpp>
